Improve condition builder
- Style problems (do not use a field, if you don't use label or something else)
- Use basic input instead of input
- Do not leak provider outside of component
- Restructure and rename components
