+++ packages/components/src/index.ts
@@ -41,7 +41,7 @@ export * from './components/common/tooltip/tooltip';
 /* Editor Components */
 export * from './components/editor/browser/browser';
 export * from './components/editor/condition-builder/conditionBuilder';
-export * from './components/editor/condition-builder/conditionBuilderContext';
+export * from './components/editor/condition-builder/conditionContext';
 export * from './components/editor/sidebar/header/header';
 export * from './components/editor/sidebar/outline/outline';
 export * from './components/editor/toolbar/toolbar';
